+++ Source/OxyPlot/Series/BarSeries/TornadoBarSeries.cs
@@ -22,6 +22,11 @@ namespace OxyPlot.Series
     public class TornadoBarSeries : CategorizedSeries
     {
         /// <summary>
+        /// The default tracker format string
+        /// </summary>
+        public new const string DefaultTrackerFormatString = "{0}\n{1}: {2}\n{3}: {4}";
+
+        /// <summary>
         /// The default fill color.
         /// </summary>
         private OxyColor defaultMaximumFillColor;
@@ -46,7 +51,7 @@ namespace OxyPlot.Series
             this.StrokeThickness = 1;
             this.BarWidth = 1;
 
-            this.TrackerFormatString = "{0}\n{1}: {2}\n{3}: {4}";
+            this.TrackerFormatString = DefaultTrackerFormatString;
             this.LabelMargin = 4;
 
             this.MinimumLabelFormatString = "{0}";