+++ CHANGELOG.md
@@ -15,6 +15,7 @@ All notable changes to this project will be documented in this file.
 - OxyPlot.Metro project (#241)
 
 ### Fixed
+- Fix exception for default tracker format strings (#265)
 - Fix Markdown links to tag comparison URL with footnote-style links.
 - WPF dispatcher issue (#311,#309)
 - Custom colors for scatters (#307)